requirements.txt
flask
openai==0.28.0
python-pptx